Refactor SecurityController and tests, create OAuthService
