+++ app/Resources/views/Default/oauth_form.html.twig
@@ -50,7 +50,7 @@
 
      <div id="status">
     </div>
-
-    <button id="_submit_oauth" name="_submit" class="hidden" />
-
 </form>
+
+<button id="btn-facebook-testhook" class="hidden" onclick="handleFacebookUserInfoResponseTrigger();"></button>
+<button id="btn-gplus-testhook" class="hidden" onclick="checkGoogleCallbackDataWithServer();"></button>
