+++ app/config/config_test.yml
@@ -1,5 +1,6 @@
 imports:
     - { resource: config_dev.yml }
+    - { resource: parameters_test.yml }
     - { resource: services_test.yml }
 
 framework:
