+++ app/config/parameters.yml.dist
@@ -28,7 +28,11 @@ parameters:
     facebook_secret: ~
     google_testuser_mail: ~
     google_testuser_pw: ~
+    google_testuser_name: ~
+    google_testuser_id: ~
     facebook_testuser_mail: ~
+    facebook_testuser_name: ~
+    facebook_testuser_id: ~
     facebook_testuser_pw: ~
 
     ldap_host: http://localhost
