+++ app/config/parameters_test.yml.dist
@@ -1,0 +1,2 @@
+parameters:
+    oauth_use_real_service: 0
