+++ src/Catrobat/AppBundle/Features/Web/oauth.feature
@@ -8,7 +8,7 @@ Feature: Open Authentication
       | Catrobat        | 123456    | cccccccccc | dev1@pocketcode.org             |
       | AlreadyinDB     | 642135    | cccccccccc | dev2@pocketcode.org             |
 
-  @javascript @insulated
+  @javascript @insulated @RealOAuth
   Scenario: Login with a new user into Facebook, logout and login again with the now existing user
     Given I am on homepage
     When I trigger Facebook login with auth_type 'reauthenticate'
@@ -28,7 +28,7 @@ Feature: Open Authentication
     And I wait for a second
     Then I should be logged in
 
-  @javascript @insulated
+  @javascript @insulated @RealOAuth
   Scenario: Login with a new user into Google+, logout and login again with the now existing user
     Given I am on homepage
     When I trigger Google login with approval prompt "force"
@@ -48,63 +48,24 @@ Feature: Open Authentication
     And I wait for a second
     Then I should be logged in
 
-  @javascript @insulated
+  @javascript
   Scenario: Try to login with a new user into Facebook where the username already exists
-    Given I am on homepage
-    When I trigger Facebook login with auth_type 'reauthenticate'
-    And I click Facebook login link
-    And I switch to popup window
-    Then I log in to Facebook with valid credentials
-    And I choose the username 'AlreadyinDB'
+    Given I am on "/pocketcode/login"
+    When I log in to Facebook with valid credentials
+    And I choose the username 'AlreadyinDB' and check button activations
     Then I should see "Username already taken, please choose a different one."
 
-  @javascript @insulated
+  @javascript
   Scenario: Try to login with a new user into Google+ where the username already exists
-    Given I am on homepage
-    When I trigger Google login with approval prompt "force"
-    And I click Google login link "twice"
-    And I switch to popup window
-    Then I log in to Google with valid credentials
+    Given I am on "/pocketcode/login"
+    When I log in to Google with valid credentials
     And I choose the username 'AlreadyinDB'
     Then I should see "Username already taken, please choose a different one."
 
-  @javascript @insulated
-  Scenario: As an OAuth user it should not be possible to change the password on the profile page or to reset the password
-    Given I am on homepage
-    When I trigger Facebook login with auth_type 'reauthenticate'
-    And I click Facebook login link
-    And I switch to popup window
-    Then I log in to Facebook with valid credentials
-    And I choose the username 'HeyWickieHey'
-    Then I should be logged in
-    And there is a user in the database:
-      | name              | email                              | facebook_uid      | facebook_name | google_uid             | google_name        |country |
-      | HeyWickieHey      | pocket_zlxacqt_tester@tfbnw.net    | 105678789764016   |               |                        |                    | en_US  |
-    And I should see an "#btn-logout" element
-    And I should see an "#btn-profile" element
-    When I click the "profile" button
-    And I wait for a second
-    Then I should not see an "#password" element
-    And I should not see an "#repeat-password" element
-    And I should see an "#btn-logout" element
-    And I click the "logout" button
-    Then I should not be logged in
-    When I trigger Facebook login with auth_type ''
-    And I click the "forgot pw or username" button
-    And I wait for a second
-    Then I should see an "#username" element
-    When I fill in "username" with "HeyWickieHey"
-    And I press "reset_pw"
-    And I wait for a second
-    Then I should see "Facebook and Google+ users do not have a password to reset."
-
-  @javascript @insulated
+  @javascript
   Scenario: It should be possible to change the E-Mail address on the profile page and login again with the same Facebook account
-    Given I am on homepage
-    When I trigger Facebook login with auth_type 'reauthenticate'
-    And I click Facebook login link
-    And I switch to popup window
-    Then I log in to Facebook with valid credentials
+    Given I am on "/pocketcode/login"
+    When I log in to Facebook with valid credentials
     And I choose the username 'HeyWickieHey'
     Then I should be logged in
     And there is a user in the database:
@@ -129,13 +90,10 @@ Feature: Open Authentication
     And I wait for a second
     Then the "email" field should contain "pocket_tester@tfbnw.net"
 
-  @javascript @insulated
+  @javascript
   Scenario: It should be possible to change the E-Mail address on the profile page and login again with the same Google+ account
-    Given I am on homepage
-    When I trigger Google login with approval prompt "force"
-    And I click Google login link "twice"
-    And I switch to popup window
-    Then I log in to Google with valid credentials
+    Given I am on "/pocketcode/login"
+    When I log in to Google with valid credentials
     And I choose the username 'PocketGoogler'
     Then I should be logged in
     And there is a user in the database:
